+++ static/js/custom_course.js
@@ -1,12 +1,15 @@
 import Vue from "vue";
+import store from "./store.js";
 import "./base.js";
 import "../css/custom_course.css";
+import AlertToast from "../../components/AlertToast.vue";
 const axios = require("axios");
 
 document.addEventListener("DOMContentLoaded", function() {
   new Vue({
     el: "#app",
     delimiters: ["[[", "]]"],
+    components: { alerttoast: AlertToast },
     data: function() {
       return {
         navBtn: false,
@@ -26,12 +29,13 @@ document.addEventListener("DOMContentLoaded", function() {
         this.computing = true;
         axios({
           method: "POST",
-          url: Flask.url_for("calendar.add_custom_course"),
+          url: Flask.url_for("custom_course.index"),
           data: evt,
           header: { "Content-Type": "application/json" }
         })
           .then(resp => {
-            console.log(resp);
+            // window.location.href = Flask.url_for("calendar.index"); // redirecting
+            store.success("Your course has been created.");
           })
           .catch(err => {
             store.error(err.response.data);
